lambda-service/lambda/requirements.txt
#transformers
#transformers[sentencepiece]
#sentencepiece
#langdetect
#datasets
#accelerate
googletrans==3.1.0a0
boto3
line-bot-sdk